student_management/backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.11.4
django-cors-headers==3.10.0
gunicorn==20.1.0
